+++ lib/app_widget.dart
@@ -13,4 +13,3 @@ class AppWidget extends StatelessWidget {
     );
   }
 }
-
