+++ lib/core/constants/sizes.dart
@@ -1,0 +1,9 @@
+class AppSizes{
+  static final int textDisplayLarge = 30;
+  static final int textDisplayMedium = 24;
+  static final int textDisplaySmall = 24;
+  static final int textBodyLarge = 16;
+  static final int textBodyMedium = 14;
+  static final int textBodySmall = 12;
+  static final int textLabelMedium = 10;
+}
