+++ lib/core/modules/app_module.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:pokedex_app/core/constants/route_names.dart';
 import 'package:pokedex_app/core/modules/home/home_module.dart';
-import 'package:pokedex_app/core/modules/splash_and_auth/auth_module/auth_module.dart';
-import 'package:pokedex_app/core/modules/splash_and_auth/splash_and_onboarding/splash_module.dart';
+import 'package:pokedex_app/core/modules/splash_and_onboarding/splash_module.dart';
 import 'package:pokedex_app/widgets/loading_screen.dart';
+import 'auth/auth_module.dart';
 
 class AppModule extends Module{
