--- lib/core/modules/splash_and_auth/splash_and_onboarding/splash.dart
@@ -1,11 +1,0 @@
-import 'package:flutter_modular/flutter_modular.dart';
-
-class Splash extends Module{
-
-
-  // @override
-  // // TODO: implement imports
-  // List<Module> get imports => [
-  //   SplashModule()
-  // ];
-}
