+++ lib/views/auth/login/login_view.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+import 'package:pokedex_app/core/constants/images.dart';
+import 'package:pokedex_app/core/constants/navigation_routes.dart';
+
+import '../widets/auth_access_view.dart';
+
+class LoginView extends StatelessWidget {
+  const LoginView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AuthAccessView(
+      appBarTitle: "Entrar",
+      presentationTitle: "Que bom te ver aqui novamente!",
+      presentationDescription: "Como deseja se conectar?",
+      presentationImage: Images.imagesLoginImage,
+      onPressedApple: () => Modular.to.navigate(NavigationRoutes.loginSuccessful),
+      onPressedGoogle: () => Modular.to.navigate(NavigationRoutes.loginSuccessful),
+      onPressedEmail: () => Modular.to.pushNamed(NavigationRoutes.loginEmail),
+    );
+  }
+}
