+++ lib/views/onboarding/onboarding_view.dart
@@ -1,0 +1,78 @@
+import 'package:flutter/material.dart';
+import 'package:carousel_slider/carousel_slider.dart';
+import 'package:flutter_modular/flutter_modular.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:pokedex_app/core/constants/colors/main_colors_light.dart';
+import 'package:pokedex_app/core/constants/images.dart';
+import 'package:pokedex_app/core/constants/navigation_routes.dart';
+import 'package:pokedex_app/core/constants/route_names.dart';
+import 'package:pokedex_app/viewmodel/onboarding/onboarding_view_model.dart';
+import 'package:pokedex_app/widgets/main_extended_button_blue.dart';
+import 'package:pokedex_app/widgets/presentation_view.dart';
+
+class OnboardingScreen extends StatefulWidget {
+  const OnboardingScreen({super.key});
+
+  @override
+  _OnboardingScreenState createState() => _OnboardingScreenState();
+}
+
+class _OnboardingScreenState extends State<OnboardingScreen> {
+  final OnboardingViewModel viewModel = Modular.get<OnboardingViewModel>();
+
+  CarouselSliderController controller = CarouselSliderController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 18),
+        child: Column(
+          children: [
+            Expanded(
+              child: CarouselSlider(
+                items: viewModel.pages,
+                carouselController: controller,
+                options: CarouselOptions(
+                  height: double.infinity,
+                  viewportFraction: 1,
+                  enableInfiniteScroll: false,
+                  onPageChanged: (index, reason) {
+                    setState(() {
+                      viewModel.onPageChanged(index);
+                    });
+                  },
+                ),
+              ),
+            ),
+            Padding(
+              padding: EdgeInsets.only(bottom: 25),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: List.generate(viewModel.pages.length, (index) {
+                  return AnimatedContainer(
+                    duration: Duration(milliseconds: 300),
+                    margin: EdgeInsets.symmetric(horizontal: 5),
+                    width: viewModel.currentPage == index ? 25 : 8,
+                    height: 8,
+                    decoration: BoxDecoration(
+                      color: viewModel.currentPage == index
+                          ? MainColorsLight.primaryBlue
+                          : Color(0x804565b7),
+                      borderRadius: BorderRadius.circular(5),
+                    ),
+                  );
+                }),
+              ),
+            ),
+
+            // Botão de Avançar
+            MainExtendedButtonBlue(
+                text: viewModel.buttonText,
+                onPressed: () => viewModel.buttonNextPage(controller))
+          ],
+        ),
+      ),
+    );
+  }
+}
